pyproject.toml
[tool.poetry]
name = "heating-emissions"
version = "1.1.1"
description = "This indicator computes annual carbon dioxide emissions from residential space heating in an area of interest within Germany"
authors = ["Sebastian Block <sebastian.block@heigit.org>", "Veit Ulrich <veit.ulrich@heigit.org>", "Rizwan Khan <Rizwan.Khan@heigit.org>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "~3.11.5"
climatoology = { git = "ssh://git@gitlab.heigit.org:2022/climate-action/climatoology.git", tag = "6.4.4" }
pydantic = "^2.8.2"
rioxarray = "^0.18.1"
rasterio = "^1.4.3"
matplotlib = "^3.10.0"
sqlalchemy = "^2.0.40"
geoalchemy2 = "^0.17.1"
scipy = "1.16.3"
cdsapi = "0.7.7"
netCDF4 = "1.7.3"
ecmwf-datastores-client = "0.4.0"

[tool.poetry.group.test.dependencies]
pytest = "^8.3.3"
pytest-cov = "^7.0.0"
pytest-mock = "^3.15.1"
pytest-postgresql = "^7.0.2"
psycopg = {extras = ["binary"], version = "^3.2.12"}
requests-mock = "^1.12.1"

[tool.poetry.group.dev.dependencies]
pre-commit = "^4.0.1"

[tool.ruff]
line-length = 120
exclude = ["heating_emissions/components/temporal_downscale/demand_ninja"]

[tool.ruff.format]
quote-style = "single"

[tool.ruff.lint]
# https://docs.astral.sh/ruff/rules
select=[
    # defaults
    "F",
    "E4", "E7", "E9",

    # additional
    "A", # be sure not to shadow built-in functions
    "I", # isort to sort imports
    "N", # check for use of PEP8 naming conventions

    # docstring cleaning
    "D201", "D207", "D208", "D209", "D210", "D214", "D215", "D300", "D403",

    # auto-remove unused 'noqa' comments
    "RUF100", "RUF101",
]

[tool.coverage.run]
omit = ["test/*" , "heating_emissions/core/*"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
